refactor: #159 - 각 API의 방어로직을 Service 레이어로 이동 - 방어로직의 경우 API에 있는 것보다 Service 클래스에 속해있는 것이 좀 더 단일 책임 원칙(SRP)에 준수한다고 판단

+++ src/main/java/freshtrash/freshtrashbackend/service/AuctionService.java
@@ -5,6 +5,7 @@ import freshtrash.freshtrashbackend.dto.request.AuctionRequest;
 import freshtrash.freshtrashbackend.dto.response.AuctionResponse;
 import freshtrash.freshtrashbackend.dto.security.MemberPrincipal;
 import freshtrash.freshtrashbackend.entity.Auction;
+import freshtrash.freshtrashbackend.entity.constants.UserRole;
 import freshtrash.freshtrashbackend.exception.AuctionException;
 import freshtrash.freshtrashbackend.exception.constants.ErrorCode;
 import freshtrash.freshtrashbackend.repository.AuctionRepository;
@@ -42,11 +43,13 @@ public class AuctionService {
                 .orElseThrow(() -> new AuctionException(ErrorCode.NOT_FOUND_AUCTION));
     }
 
-    public void deleteAuction(Long auctionId) {
+    public void deleteAuction(Long auctionId, UserRole userRole, Long memberId) {
+        checkIfWriterOrAdmin(auctionId, userRole, memberId);
         auctionRepository.deleteById(auctionId);
     }
 
-    public boolean isWriterOfAuction(Long auctionId, Long memberId) {
-        return auctionRepository.existsByIdAndMemberId(auctionId, memberId);
+    private void checkIfWriterOrAdmin(Long auctionId, UserRole userRole, Long memberId) {
+        if (userRole != UserRole.ADMIN && !auctionRepository.existsByIdAndMemberId(auctionId, memberId))
+            throw new AuctionException(ErrorCode.FORBIDDEN_AUCTION);
     }
 }
